Higher order component for private route updated and loader added
