+++ apps/goals-on-track/component/common/layout/layout.tsx
@@ -1,6 +1,7 @@
 import styles from './layout.module.scss';
 import { Header , Footer } from '@goal-tracker/ui';
 // import { Header } from '@goal-tracker/ui'
+import { useRouter } from 'next/router';
 
 /* eslint-disable-next-line */
 export interface LayoutProps {
@@ -8,6 +9,9 @@ export interface LayoutProps {
 }
 
 export function Layout(props: LayoutProps) {
+  const router = useRouter()
+  console.log('x', router.pathname.includes('/'))
+
   return (
     <div className={styles.container}>
       <Header />
