+++ libs/shared/ui/src/lib/components/Loader/Loader.module.scss
@@ -1,0 +1,25 @@
+.loader_container{
+    width: 100vw;
+    height: 100vh;
+    background-color: white;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    flex-direction: column;
+    font-size: 8vw;
+    font-weight: 600;
+    color: '#232f3e';
+}
+.loader_container div{
+    width: 8vmax;
+    height: 8vmax;
+    border-bottom: 5px solid #232f3e;
+    border-radius: 50%;
+    animation: loader_container_Rotate 800ms linear infinite;
+}
+
+@keyframes loader_container_Rotate {
+    to{
+        transform: rotateZ(360deg);
+    }
+}
