+++ libs/shared/ui/src/lib/components/Loader/Loader.spec.tsx
@@ -1,0 +1,10 @@
+import { render } from '@testing-library/react';
+
+import Loader from './Loader';
+
+describe('Loader', () => {
+  it('should render successfully', () => {
+    const { baseElement } = render(<Loader />);
+    expect(baseElement).toBeTruthy();
+  });
+});
