+++ libs/shared/ui/src/lib/components/Login/Login.spec.tsx
@@ -1,0 +1,10 @@
+import { render } from '@testing-library/react';
+
+import Login from './Login';
+
+describe('Login', () => {
+  it('should render successfully', () => {
+    const { baseElement } = render(<Login />);
+    expect(baseElement).toBeTruthy();
+  });
+});
